unify transfer server, fix port, fix persistance, fix race condition in multitenancy
